Add files via upload
SQL query to retrieve the checkNumber, paymentDate, and amount from the payments table.
SQL query to retrieve the orderDate, requiredDate, and status of orders that are currently 'In Process' from the orders table. Sortting the results in descending order of orderDate
SQL query to display the firstName, lastName, and email of employees whose job title is 'Sales Rep' and order them in descending order of employeeNumber.
SQL query to retrieve all the columns and records from the offices table.
SQL query to fetch the productName and quantityInStock from the products table. Sort the results in ascending order of buyPrice and limit the output to 5 records.